Resolving pr comments for service account count
