+++ src/Core/Repositories/IOrganizationUserRepository.cs
@@ -41,5 +41,4 @@ public interface IOrganizationUserRepository : IRepository<OrganizationUser, Gui
     Task RestoreAsync(Guid id, OrganizationUserStatusType status);
     Task<IEnumerable<OrganizationUserPolicyDetails>> GetByUserIdWithPolicyDetailsAsync(Guid userId, PolicyType policyType);
     Task<int> GetOccupiedSmSeatCountByOrganizationIdAsync(Guid organizationId);
-    Task<int> GetOccupiedServiceAccountCountByOrganizationIdAsync(Guid organizationId);
 }
