+++ src/Core/SecretsManager/Repositories/IServiceAccountRepository.cs
@@ -15,4 +15,5 @@ public interface IServiceAccountRepository
     Task<bool> UserHasWriteAccessToServiceAccount(Guid id, Guid userId);
     Task<IEnumerable<ServiceAccount>> GetManyByOrganizationIdWriteAccessAsync(Guid organizationId, Guid userId, AccessClientType accessType);
     Task<(bool Read, bool Write)> AccessToServiceAccountAsync(Guid id, Guid userId, AccessClientType accessType);
+    Task<int> GetServiceAccountCountByOrganizationIdAsync(Guid organizationId);
 }
