+++ src/Infrastructure.EntityFramework/Repositories/OrganizationUserRepository.cs
@@ -628,14 +628,4 @@ public class OrganizationUserRepository : Repository<Core.Entities.OrganizationU
         return await GetCountFromQuery(query);
     }
 
-    public async Task<int> GetOccupiedServiceAccountCountByOrganizationIdAsync(Guid organizationId)
-    {
-        using (var scope = ServiceScopeFactory.CreateScope())
-        {
-            var dbContext = GetDatabaseContext(scope);
-            return await dbContext.ServiceAccount
-                .Where(ou => ou.OrganizationId == organizationId)
-                .CountAsync();
-        }
-    }
 }
